working pretty well

+++ src/server.c
@@ -6,7 +6,7 @@
 /*   By: alorain <alorain@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/06 15:33:32 by alorain           #+#    #+#             */
-/*   Updated: 2022/01/07 20:12:46 by alorain          ###   ########.fr       */
+/*   Updated: 2022/01/08 16:36:36 by alorain          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,16 +21,14 @@ int	acknowledgement(int c_pid, int idx)
 
 	i = idx;
 	count = 0;
-	if (i % 8 == 0)
+	if ( i && i % 8 == 0)
 	{
-		while (i > idx - 8)
+		while (i > idx - 8 - 1)
 		{
 			if (g_talk.buffer[i] == '0')
 				count++;
 			i--;
 		}
- 		/* printf("idx = %d\n", idx);
- 		printf("count = %d\n", count); */
 		if (count == 8)
 		{
 			if(kill(c_pid, SIGUSR1) == -1)
@@ -48,20 +46,11 @@ void	convert_buff(char *str, int idx)
 
 	idx_str = 0;
 	i = 0;
-	while (i < idx)
+	while (i <= idx)
 	{
 		if (i && i % 8 == 0)
 		{
-			str[i] = ft_bitoa(&g_talk.buffer[i - 8]);
-			ft_printf("char =%c\n", ft_bitoa(&g_talk.buffer[i - 8]));
-			/* int index;
-
-			index =  0;
-			while (index <= idx)
-			{ */
-				ft_printf("str =%c\n",g_talk.buffer[/* index */0]);
-			/* 	index++;
-			} */
+			str[idx_str] = ft_bitoa(&g_talk.buffer[i - 8]);
 			idx_str++;
 		}
 		i++;
@@ -69,20 +58,56 @@ void	convert_buff(char *str, int idx)
 	str[idx_str] = '\0';
 }
 
+void	ft_strlcat2(char *s1, char *s2)
+{
+	int i;
+	int j;
+
+	i = 0;
+	j = 0;
+	while (s1[i])
+		i++;
+	while (s2[j])
+	{
+		s1[i + j] = s2[j];
+		j++;
+	}
+	s1[i+j] = 0;
+}
+
+void	concat(char *str)
+{
+	int		old_len;
+	int		new_len;
+	char	*new_str;
+
+	old_len = ft_strlen(g_talk.str) + 1;
+	new_len = old_len + ft_strlen(str);
+	/* ft_printf("new len =%d\n", new_len); */
+	new_str = malloc(sizeof(char) * (new_len + 1));
+	new_str[0] = 0;
+	ft_strlcat2(new_str, g_talk.str/* , old_len */);
+	ft_strlcat2(new_str, str/* , ft_strlen(str) */);
+	new_str[new_len + 1] = 0;
+	/* ft_printf("str = %s\n", new_str); */
+	free(g_talk.str);
+	g_talk.str = new_str;
+}
+
 void	manage_buffer(int idx)
 {
 	char	*str;
 
-	str = malloc(sizeof(char) * (idx / 8));
+	str = malloc(sizeof(char) * (idx / 8) /* 32 */);
 	convert_buff(str, idx);
-	ft_printf("sizeof str = %d\n", idx / 8);
-	ft_printf("buffer str = %s\n", str );
-	g_talk.str = str;
+	concat(str);
+	free(str);
 }
 
 void	flush_str()
 {
 	ft_printf("%s\n", g_talk.str);
+	g_talk.str[0] = 0;
 }
 
 void	sig_pid(int sig, siginfo_t *info, void *content)
@@ -97,19 +122,23 @@ void	sig_pid(int sig, siginfo_t *info, void *content)
 		if (sig == SIGUSR1)
 		{
 			g_talk.buffer[idx] = '1';
-			ft_printf("1");
+			/* ft_printf("%c", g_talk.buffer[idx]); */
 		}
 		else if (sig == SIGUSR2)
 		{
 			g_talk.buffer[idx] = '0';
-			ft_printf("0");
+			/* ft_printf("%c", g_talk.buffer[idx]); */
 		}
+		/* if (idx && (idx + 1) % 8 == 0)
+			ft_printf(" "); */
 		idx++;
-		usleep(600);
-		/* g_talk.buffer[idx + 1] = 0; */
+		usleep(100);
+		g_talk.buffer[idx + 1] = 0;
+		kill(c_pid, SIGUSR2);
 	}
 	if (idx >= BUFFER_SIZE)
 	{
+		/* ft_printf("yo"); */
 		manage_buffer(idx);
 	}
 	if (acknowledgement(c_pid, idx))
@@ -118,8 +147,8 @@ void	sig_pid(int sig, siginfo_t *info, void *content)
 		flush_str();
 		idx = 0;
 	}
-	else
-		kill(c_pid, SIGUSR2);
+	else if (idx >= BUFFER_SIZE)
+		idx = 0;
 }
 
 int main()
@@ -127,7 +156,7 @@ int main()
 	struct sigaction	sa;
 
 	g_talk.str = malloc(sizeof(char));
-	g_talk.str[0] = '\0'; 
+	g_talk.str[0] = '\0';
 	ft_printf("PID: %d\n", getpid());
 	sigemptyset(&sa.sa_mask);
 	sa.sa_flags = SA_SIGINFO;
